% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BICadj.R
\name{BICadj}
\alias{AICadj}
\alias{BICadj}
\alias{varexp}
\title{Ways to compare SITAR models for fit}
\usage{
BICadj(..., pattern = NULL)

AICadj(..., k = 2, pattern = NULL)

varexp(..., pattern = NULL)
}
\arguments{
\item{pattern}{regular expression defining names of SITAR models.}

\item{k}{numeric, the penalty per parameter to be used; the default k = 2 is
the classical AIC.}

\item{\dots}{one or more SITAR models.}
}
\value{
For \code{BICadj} and \code{AICadj} a named vector of deviances in
increasing order.  For \code{varexp} a named vector of percentages in
decreasing order.
}
\description{
\code{BICadj} and \code{AICadj} calculate the BIC and AIC for SITAR models,
adjusting the likelihood for Box-Cox transformed y variables. \code{varexp}
calculates the variance explained by SITAR models, compared to the
corresponding fixed effect models.
}
\examples{
data(heights)
## fit sitar model for height
m1 <- sitar(x=age, y=height, id=id, data=heights, df=5)

## update it for log(height)
m2 <- update(m1, y=sqrt(height))

## compare variance explained in the two models
varexp(m1, m2)

## compare BIC adjusting for sqrt transform
## the pattern matches names starting with "m" followed by a digit
BICadj(pattern="^m[0-9]")

}
\author{
Tim Cole \email{tim.cole@ucl.ac.uk}
}
\seealso{
\code{\link{BIC}}, \code{\link{AIC}}
}
\keyword{regression}

