% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfpower.R
\name{dfpower}
\alias{dfpower}
\title{Tabulate BIC of SITAR models by degrees of freedom and xy power transformations}
\usage{
dfpower(model, df, xpowers = 1, ypowers = xpowers,
  maxIter = nlmeControl()$maxIter, verbose = FALSE)
}
\arguments{
\item{model}{fitted \pkg{sitar} model to be updated.}

\item{df}{vector of degrees of freedom to be fitted.}

\item{xpowers}{vector of powers to apply to x.}

\item{ypowers}{vector of powers to apply to y.}

\item{maxIter}{maxium number of iterations.}

\item{verbose}{logical controlling monitoring.}
}
\value{
3-way table of BIC by df, x and y powers.
}
\description{
\code{dfpower} fits a series of SITAR models for specified degrees of freedom
and power transformations of x and y, returning a table of adjusted BIC.
}
\details{
The function provides a convenient way to optimise the model's degrees of freedom
and explore transformations of x and y.
Adjusted BIC is obtained using \code{\link{BICadj}}, and
is set negative for models failing to converge; the
run-time can be shortened by reducing \code{maxIter} appropriately.
The returned table can be rearranged using \code{\link{aperm}}.
}
\examples{
data(heights)
m1 <- sitar(log(age), height, id, heights, 4)
dfpower(m1, 4:6, 1:0)
}
\author{
Tim Cole \email{tim.cole@ucl.ac.uk}
}

