\name{getData.sitar}
\alias{getData.sitar}
\title{Extract SITAR object data}
\description{
getData method for \code{sitar} objects, based on \code{getData.nlme}.}
\usage{\method{getData}{sitar}(object)}
\arguments{
  \item{object}{
an object inheriting from class \code{sitar}.}
}
\value{The data frame used in the fit.}
\author{Tim Cole \email{tim.cole@ucl.ac.uk}}
