% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeakTrough.R
\name{getPeakTrough}
\alias{getPeakTrough}
\alias{getPeak}
\alias{getTrough}
\title{Identify peak or trough of curve}
\usage{
getPeakTrough(x, y = NULL, peak = TRUE)

getPeak(x, y = NULL, peak = TRUE)

getTrough(x, y = NULL, peak = FALSE)
}
\arguments{
\item{x}{vector.}

\item{y}{vector.}

\item{peak}{logical determining whether peak or trough is returned.}
}
\value{
A length-2 vector containing the values of \code{x} and  \code{y}
at the peak or trough. If no peak/trough is identified \code{x} and  \code{y}
are set to NA.
}
\description{
Given vectors \code{x} and  \code{y}, returns their values at the (highest)
peak or (lowest) trough of the curve \code{y ~ x}, where \code{Dy} is zero and
\code{D2y} is negative or positive respectively.
}
\details{
Note that turning points where the curvature is close to zero (and the curve
close to linear) are ignored.
}
\examples{
## create mean height velocity curve
data(heights)
m1 <- sitar(age, height, id, heights, 4)
xy <- plot_v(m1)
## and plot it
plot(xy, type='l', xlab='age', ylab='height velocity')
points(t(getPeak(xy)), pch=17)
points(t(getTrough(xy)), pch=25, bg=1)
}
\author{
Tim Cole \email{tim.cole@ucl.ac.uk}
}
