\name{getVarCov.sitar}
\alias{getVarCov.sitar}
\title{Extract variance-covariance matrix for SITAR models}
\description{
getVarCov method for \code{sitar} objects, based on \code{getVarCov.lme}.}
\usage{\method{getVarCov}{sitar}(obj, ...)}
\arguments{
  \item{obj}{
an object inheriting from class \code{sitar}.}
  \item{\dots}{
other optional arguments.}
}
\value{The variance-covariance matrix.}
\author{Tim Cole \email{tim.cole@ucl.ac.uk}}
