% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ob_convertr.R
\name{ob_convertr}
\alias{ob_convertr}
\title{Convert between IOTF, WHO and CDC prevalence rates for child underweight, overweight and obesity}
\usage{
ob_convertr(
  prev = 50,
  age,
  sex,
  from,
  to,
  prev_true = NA,
  report = c("vector", "wider", "longer"),
  plot = c("no", "density", "compare"),
  data = parent.frame()
)
}
\arguments{
\item{prev}{vector of percentage prevalence rates in groups of children.}

\item{age}{vector of mean ages in years corresponding to each rate.}

\item{sex}{vector of the sexes corresponding to each rate, coded as either 'boys/girls'
or 'male/female' or '1/2' (only the first character in each case counts).}

\item{from}{the BMI cutoff (see below) on which prevalence \code{prev} of
underweight, overweight or obesity is based.}

\item{to}{the BMI cutoff to be used to convert the prevalence to.}

\item{prev_true}{optional vector of known prevalence rates corresponding to cutoff \code{to},
for comparison purposes.}

\item{report}{character controlling the format of the returned data: 'vector'
for the estimated prevalence rates, 'wider' for the working tibble in wide format,
i.e. the 'from' and 'to' data side by side, and 'longer' for the tibble in
long format, i.e. two rows, one for 'from' and one for 'to'.}

\item{plot}{character controlling what if anything is plotted: 'no' for no
plot, 'density' to compare the BMI distributions and cutoffs corresponding
to \code{from} and \code{to}, or 'compare' to compare the estimated prevalence
rates with the true rates in \code{prev_true}.}

\item{data}{data frame containing prev, age, sex and optionally prev_true.}
}
\value{
Either the estimated prevalence rates based on cutoff \code{to}, or a plot
visualizing the findings, depending on the \code{report} and \code{plot} settings.
}
\description{
Child thinness, overweight and obesity are defined by the child's body mass index (BMI) exceeding
(above or below) a pre-specified reference cutoff. Three references are considered:
IOTF (International Obesity Task Force), WHO (World Health Organization) and CDC
(US Centers for Disease Control and Prevention). They each have separate cutoffs
for thinness, overweight and obesity. \code{ob_convertr} takes age-sex-specific
prevalence rates based on one reference cutoff and estimates what the rates would be
if based on a different cutoff, using a novel algorithm.
}
\details{
Each reference has cutoffs corresponding to respectively underweight, overweight and obesity.
They are referred to here as "IOTF16", "IOTF17", "IOTF18.5", "IOTF25", "IOTF30", "IOTF35",
"WHO-2", "WHO-1", "WHO+1", "WHO+2", "CDC5", "CDC85" and "CDC95".

See ... for an explanation of how the conversion calculation is done.
}
\examples{
## convert 10\% IOTF overweight prevalence (cutoff IOTF25) in 8-year-old boys
## to the overweight prevalence based on WHO, i.e. cutoff WHO+1
ob_convertr(prev = 10, age = 8, sex = 'boys', from = 'IOTF25', to = 'WHO+1')

## compare the BMI density functions and cutoffs for IOTF25 and WHO+1
ob_convertr(10, 8, 'boys', from = 'IOTF25', to = 'WHO+1', plot = 'density')
}
\author{
Tim Cole \email{tim.cole@ucl.ac.uk}
}
