% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.sitar.R
\name{anova.sitar}
\alias{anova.sitar}
\title{Compare Likelihoods of Fitted SITAR Objects}
\usage{
\method{anova}{sitar}(object, ..., test = TRUE, type = c("sequential",
  "marginal"), adjustSigma = TRUE, Terms, L, verbose = FALSE)
}
\arguments{
\item{object}{an object inheriting from class \code{sitar}.}

\item{test}{an optional logical value controlling whether likelihood ratio
tests should be used.}

\item{type}{an optional character string specifying the type of sum of
squares to be used.}

\item{adjustSigma}{see \code{\link{anova.lme}}.}

\item{Terms}{see \code{\link{anova.lme}}.}

\item{L}{see \code{\link{anova.lme}}.}

\item{verbose}{an optional logical value.}

\item{\dots}{other optional fitted model objects.}
}
\value{
a data frame inheriting from class "anova.lme".
}
\description{
anova method for \code{sitar} objects, based on \code{anova.lme}.
}
\author{
Tim Cole \email{tim.cole@ucl.ac.uk}
}

