% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitarlib.R
\name{funcall}
\alias{funcall}
\title{Function call with optional inverse}
\usage{
funcall(v, vcall, inverse = FALSE)
}
\arguments{
\item{v}{vector}

\item{vcall}{expression}

\item{inverse}{logical}
}
\value{
Returns a vector of length v.
}
\description{
Applies an expression to vector v, optionally inverting the expression
first. For example if the expression is log, funcall returns log(v) if
inverse is FALSE, and exp(v) if inverse is TRUE.
}
\details{
Inverse covers functions log, exp, sqrt, ^, *, /, +, -.
}
\author{
Tim Cole \email{tim.cole@ucl.ac.uk}
}

