\name{getCovariate.sitar}
\alias{getCovariate.sitar}
\title{Extract Covariate from a SITAR object}
\description{
getCovariate method for \code{sitar} objects.}
\usage{\method{getCovariate}{sitar}(object, ...)}
\arguments{
  \item{object}{an object inheriting from class \code{sitar}.}
  \item{\dots}{optional arguments, not used here.}
}
\value{The \code{x} variable used in the SITAR fit.}
\author{Tim Cole \email{tim.cole@ucl.ac.uk}}
