% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.sitar.R
\name{getData.sitar}
\alias{getCovariate.sitar}
\alias{getData.sitar}
\alias{getVarCov.sitar}
\title{Extract elements of fitted SITAR models}
\usage{
\method{getData}{sitar}(object)

\method{getCovariate}{sitar}(object, ...)

\method{getVarCov}{sitar}(obj, ...)
}
\arguments{
\item{object, obj}{an object inheriting from class \code{sitar}.}

\item{\dots}{other optional arguments.}
}
\value{
Respectively the data frame and \code{x} variable
used in the fit, and the returned variance-covariance matrix.
}
\description{
getData, getCovariate and getVarCov methods for \code{sitar} objects,
based on \code{lme}.
}
\author{
Tim Cole \email{tim.cole@ucl.ac.uk}
}

