% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeakTrough.R
\name{getPeakTrough}
\alias{getPeakTrough}
\alias{getPeak}
\alias{getTrough}
\title{Identify peak or trough of curve}
\usage{
getPeakTrough(x, y = NULL, peak = TRUE, use = "value")

getPeak(x, y = NULL, peak = TRUE, use = "value")

getTrough(x, y = NULL, peak = FALSE, use = "value")
}
\arguments{
\item{x}{vector.}

\item{y}{vector.}

\item{peak}{logical determining whether peak or trough is returned.}

\item{use}{character string defining the criterion to use - "value"
(default) or "curvature".}
}
\value{
A length-2 vector containing the values of \code{x} and \code{y} at
the peak or trough, defined as the maximum/minimum value or minimum/maximum
curvature, depending on \code{use}. If no peak/trough is identified NA's
are returned.
}
\description{
Given vectors \code{x} and  \code{y}, returns their values at the (highest)
peak or (lowest) trough of the smooth curve \code{y ~ x} (typically a fitted
cubic spline curve), where \code{Dy} is zero and \code{D2y} is negative or
positive respectively.
}
\examples{
## create mean height velocity curve
data(heights)
m1 <- sitar(age, height, id, heights, 4)
## plot it
plot(m1, 'v')
## mark peak and trough
xy <- plot_v(m1)
points(t(getPeak(xy)), las=1, pch=17)
points(t(getTrough(xy)), pch=25, col=2, bg=2)
## wrong trough - use curvature criterion instead
points(t(getTrough(xy, use='curvature')), pch=25, bg=1)
}
\author{
Tim Cole \email{tim.cole@ucl.ac.uk}
}
