% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeakTakeoff.R
\name{getPeakTakeoff}
\alias{getPeakTakeoff}
\alias{getPeak}
\alias{getTakeoff}
\title{Identify peak or takeoff point of velocity curve}
\usage{
getPeakTakeoff(x, y = NULL, peak = TRUE)

getPeak(x, y = NULL, peak = TRUE, use = "value")

getTakeoff(x, y = NULL, peak = FALSE, use = "value")
}
\arguments{
\item{x}{vector.}

\item{y}{vector.}

\item{peak}{logical determining whether peak or takeoff is returned.}
}
\value{
A length-2 vector containing the values of \code{x} and \code{y} at
the peak or takeoff. If no peak/takeoff is identified NA's are returned.
}
\description{
Given vectors \code{x} and  \code{y}, returns their values at the (highest)
peak or (lowest) takeoff point of the smooth (e.g. cubic spline) velocity
curve \code{y ~ x}, where \code{Dy} is zero and \code{D2y} is
negative or positive respectively. The takeoff point is required to occur
before the peak.
}
\examples{
## create mean height velocity curve
data(heights)
m1 <- sitar(age, height, id, heights, 4)
## plot it
plot(m1, 'v')
## mark peak and takeoff
xy <- plot_v(m1)
points(t(getPeak(xy)), las=1, pch=17)
points(t(getTakeoff(xy)), pch=25, col=2, bg=2)
}
\author{
Tim Cole \email{tim.cole@ucl.ac.uk}
}
