% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sitar.R
\name{plot.sitar}
\alias{lines.sitar}
\alias{plot.sitar}
\title{Plot SITAR model}
\usage{
\method{plot}{sitar}(x, opt = "dv", labels, apv = FALSE, xfun = NULL,
  yfun = NULL, subset = NULL, abc = NULL, add = FALSE, nlme = FALSE,
  ...)

\method{lines}{sitar}(x, ...)
}
\arguments{
\item{x}{object of class \code{sitar}.}

\item{opt}{character string containing a subset of letters corresponding to
the options: 'd' for fitted Distance curve, 'v' for fitted Velocity curve,
'e' for fitted fixed Effects distance curve, 'D' for individual fitted
Distance curves, 'V' for individual fitted Velocity curves, 'u' for
Unadjusted individual growth curves, and 'a' for Adjusted individual growth
curves. Options 'dveDV' give spline curves, while 'ua' give data curves made
up as line segments. If both distance and velocity curves are specified, the
axis for the velocity curve appears on the right side of the plot (y2), and
a legend identifying the distance and velocity curves is provided.}

\item{labels}{optional character vector containing plot labels for \code{x},
\code{y} and \code{y} velocity from the original SITAR model. The first two
elements can alternatively be provided via \code{\link{par}} parameters
\code{xlab} and \code{ylab}, and the third element via \code{y2par} (see
Details). Default labels are the names of \code{x} and \code{y}, and
"\code{y} velocity", suitably adjusted to reflect any back-transformation
via \code{xfun} and \code{yfun}.}

\item{apv}{optional logical specifying whether or not to calculate the age
at peak velocity from the velocity curve. If TRUE, age at peak velocity is
calculated as the age when the second derivative of the fitted curve changes
sign (after applying \code{xfun} and/or \code{yfun}). Age at peak velocity
is marked in the plot with a vertical dotted line, and its value, along with
peak velocity, is printed and returned.}

\item{xfun}{optional function to be applied to the x variable prior to
plotting. Defaults to NULL, which translates to \code{ifun(x$call.sitar$x)}
and inverts any transformation applied to x in the original SITAR model
call. To plot on the transformed scale set \code{xfun} to \code{I}.}

\item{yfun}{optional function to be applied to the y variable prior to
plotting. Defaults to NULL, which translates to \code{ifun(x$call.sitar$y)}
and inverts any transformation applied to y in the original SITAR model
call. To plot on the transformed scale set \code{yfun} to \code{I}.}

\item{subset}{optional logical vector of length \code{x} defining a subset
of \code{data} rows to be plotted.}

\item{abc}{vector of named values of random effects a, b and c used to
define an individual growth curve, e.g. abc=c(a=1, c=-0.1). Alternatively a
single character string defining an \code{id} level whose random effect
values are used. If \code{abc} is set, \code{level} is ignored. If
\code{abc} is NULL (default), or if a, b or c values are missing, values of
zero are assumed.}

\item{add}{optional logical defining if the plot is pre-existing (TRUE) or
new (FALSE). TRUE is equivalent to using \code{lines}.}

\item{nlme}{optional logical which set TRUE plots the model as an
\code{nlme} object, using \code{plot.nlme} arguments.}

\item{\dots}{Further graphical parameters (see \code{par}) may also be
supplied as arguments, e.g. axis labels \code{xlab} and \code{ylab}, line
type \code{lty}, line width \code{lwd}, and color \code{col}. For the
velocity (y2) plot \code{y2par} can be used (see Details).}
}
\value{
Returns invisibly a list of three objects:
\item{ss}{\code{smooth.spline} object corresponding to the fitted distance
curve.} \item{usr}{value of \code{par('usr')} for the main plot.}
\item{usr2}{the value of \code{par('usr')} for the velocity (y2) plot.}
\item{apv}{if argument \code{apv} is TRUE a named list giving the age at
peak velocity (apv) and peak velocity (pv) from the fitted velocity curve.}
}
\description{
\code{plot} and \code{lines} methods for objects of class \code{sitar},
providing various flavours of plot of the fitted growth curves.
}
\details{
For option 'dv' (the default) the velocity curve plot (with right axis) can
be annotated with \code{par} parameters given as a named list called
\code{y2par}. To suppress the legend that comes with it set \code{xlegend =
NULL}.
}
\examples{

## fit sitar model
m1 <- sitar(x=age, y=height, id=id, data=heights, df=7)

## draw fitted distance and velocity curves
## with velocity curve in blue
## adding age at peak velocity
plot(m1, y2par=list(col='blue'), apv=TRUE)

## draw individually coloured growth curves adjusted for random effects
## using same x-axis limits as for previous plot
plot(m1, opt='a', col=id, xlim=xaxsd())

## add mean curve in red
lines(m1, opt='d', col='red', lwd=2)

## add mean curve for a, b, c = -1 SD
lines(m1, opt='d', lwd=2, abc=-sqrt(diag(getVarCov(m1))))

}
\author{
Tim Cole \email{tim.cole@ucl.ac.uk}
}
\seealso{
\code{\link{mplot}},
\code{\link{plotclean}}, \code{\link{y2plot}}, \code{\link{ifun}}
}
\keyword{aplot}

