\name{print.sitar}
\alias{print.sitar}
\title{Print SITAR model}
\description{
Print method for \code{sitar} objects, based on \code{nlme:::print.lme}.}
\usage{\method{print}{sitar}(x, ...)}
\arguments{
  \item{x}{
an object inheriting class \code{sitar}.}
  \item{\dots}{
other optional arguments.}
}
\author{Tim Cole \email{tim.cole@ucl.ac.uk}}
