% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitarlib.R
\name{print.sitar}
\alias{print.sitar}
\title{Print SITAR model}
\usage{
\method{print}{sitar}(x, ...)
}
\arguments{
\item{x}{an object inheriting class \code{sitar}.}

\item{\dots}{other optional arguments.}
}
\description{
Print method for \code{sitar} objects, based on \code{print.lme}.
}
\author{
Tim Cole \email{tim.cole@ucl.ac.uk}
}

