\name{print.summary.sitar}
\alias{print.summary.sitar}
\title{Print summary of SITAR model}
\description{
A \code{print.summary} method for \code{sitar} objects.}
\usage{
\method{print}{summary.sitar}(x, verbose = FALSE, ...)
}
\arguments{
  \item{x}{
an object inheriting from class \code{summary.sitar}.}
  \item{verbose}{
a logical to control the amount of output.}
  \item{\dots}{
to specify extra arguments.}
}
\value{
A formatted summary of the object.
}
\author{Tim Cole \email{tim.cole@ucl.ac.uk}}
