% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitarlib.R
\name{print.summary.sitar}
\alias{print.summary.sitar}
\title{Print summary of SITAR model}
\usage{
\method{print}{summary.sitar}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{an object inheriting from class \code{summary.sitar}.}

\item{verbose}{a logical to control the amount of output.}

\item{\dots}{to specify extra arguments.}
}
\value{
A formatted summary of the object.
}
\description{
A \code{print.summary} method for \code{sitar} objects.
}
\author{
Tim Cole \email{tim.cole@ucl.ac.uk}
}
