\name{revise}
\alias{revise}
\title{revise}
\description{Change old-style \code{sitar} obect to new-style. Outdated, and better to use \code{update}.}
\usage{
revise(obj)
}
\arguments{
  \item{obj}{
an old-style \code{sitar} object.}
}
\value{
Returns an new-style \code{sitar} object.}
\author{Tim Cole \email{tim.cole@ucl.ac.uk}}
\seealso{
\code{\link{nk2df}}, \code{\link{update}}
}
